+++ frontend/src/components/AACBoard/SelectedItemsDisplay.jsx
