+++ frontend/src/components/IngredientScrollPicker/IngredientScrollPicker.css
