+++ frontend/src/components/Manager/ManagerActions.css
