+++ frontend/src/components/Manager/MiniOrderDisplay.css
