+++ frontend/src/components/Manager/MiniOrderDisplay.jsx
@@ -1,9 +1,11 @@
-import BurgerDisplay from "./BurgerDisplay";
-import SideDisplay from "./SideDisplay";
-import DrinkDisplay from "./DrinkDisplay";
-import "./MiniOrderDisplay.css"
 import React from "react";
-import { menuMap } from "../menuItems";
+import BurgerDisplay from "../Burger/BurgerDisplay";
+import SideDisplay from "../Sides/SideDisplay";
+import DrinkDisplay from "../Drinks/DrinkDisplay";
+import { menuMap } from "../../menuItems";
+
+import "./MiniOrderDisplay.css";
+
 
 export default function MiniOrderDisplay({ burger, side, drink }) {
     const isBurger = burger && burger.length > 0;
@@ -30,7 +32,7 @@ export default function MiniOrderDisplay({ burger, side, drink }) {
                     color={drink.color}
                     fillPercentage={drink.fill}
                     cupSize={drink.size}
-                    mini = {true}
+                    mini={true}
                 />
             )}
             {isSide && (<SideDisplay tableState={side.table_state} />)}
